+++ src/commands/Info/help.ts
@@ -1,5 +1,5 @@
 import { Message } from "discord.js";
-import { prefix } from "../config.json";
+import { prefix } from "../../config.json";
 import fs from "fs";
 
 module.exports = {
